Allow non-integer XP.

+++ src/persons/Person.cs
@@ -351,7 +351,7 @@ public class Person : ISkillContext, IAbilityContext, IInventoryContext, IHouseh
     }
   }
 
-  public bool GrantXP(Skill skill, int xp)
+  public bool GrantXP(Skill skill, double xp)
   {
     return skills.GrantXP(skill, xp);
   }
@@ -371,12 +371,12 @@ public class Person : ISkillContext, IAbilityContext, IInventoryContext, IHouseh
     return skills.GetLevel(skill);
   }
 
-  public int GetXP(Skill skill)
+  public double GetXP(Skill skill)
   {
     return skills.GetXP(skill);
   }
 
-  public int GetNextLevelXP(Skill skill)
+  public double GetNextLevelXP(Skill skill)
   {
     return skills.GetNextLevelXP(skill);
   }
